Various indentation fixes
Most of Planner's code uses tabs at the beginning of lines and spaces in all
other places, such as between a function name and its list of parameters.
Spaces are also used after tabs on lines that need to align with previous
lines, such as function prototypes that are split over multiple lines.
This patch fixes some code where this style was not used.