Added assigned resource information to HTML export
The HTML export now lists both in the gantt chart and in the task view
what resources are assigned to a task and for how many units if it is
not 100. The code was also cleaned up a bit to have less duplication.

The priority field was removed because it is not used anywhere and was
only visible in the HTML export and in the task dialog.