Fix crash when using zoom-to-fit (bug #550559)
The problem was caused by calling gnome_canvas_get_scroll_offsets() from
gantt_chart_reflow_idle(). The scroll offsets were used to prevent scrolling in
some cases when the duration of a task was being dragged. Unfortunately the
returned scroll offsets seemed to not always be up-to-date, eventually
resulting in a crash.

This patch uses a dummy canvas item to extend the scroll region while dragging
the duration of a task instead of relying on gnome_canvas_get_scroll_offsets().